+++ src/Type/WPObject/Order/OrderItemOption.php
@@ -14,7 +14,6 @@ use WPGraphQL\AppContext;
 use WPGraphQL\GF\Data\Loader\FormFieldsLoader;
 use WPGraphQL\GF\Type\WPInterface\FormField;
 use WPGraphQL\GF\Type\WPObject\AbstractObject;
-use WPGraphQL\GF\Utils\Compat;
 
 /**
  * Class - OrderItemOption
@@ -43,7 +42,7 @@ class OrderItemOption extends AbstractObject {
 				'type'        => FormField::$type,
 				'description' => static fn () => __( 'The form field that the order item is connected to', 'wp-graphql-gravity-forms' ),
 				'resolve'     => static function ( $source, array $args, AppContext $context ) {
-					$gf_form = Compat::get_app_context( $context, 'gfForm' );
+					$gf_form = $context->get( 'gf', 'gfForm' );
 					if ( ! isset( $gf_form ) || ! isset( $source['id'] ) ) {
 						return null;
 					}